Switch to using a single metric.
Cloudwatch expects data to always be present in order for alarms to work
as expected. Therefore we are going to switch to using a single metric
for storing successes and failures. Failures can then be detected by
their value, of 1 whereas total operations can be found by counting
the number of data points.